Move migration to run after post-configure
This change has been discussed and planned for a while. We have not used
migrations much in TripleO, so it was not clear that it was necessary.
But as we move to HA deployments we need a way to run scripts after the
system state has been asserted, which is what migration is for.

Change-Id: I16d90abb3fccc791001beb5d8ed267fcd00f820c